fix: update api variable name & folder structure
