+++ apis/search/getCategoryIngredient.ts
@@ -1,6 +1,4 @@
-import { useQuery } from '@tanstack/react-query';
 import axiosClient from '../axios';
-import { useMemo } from 'react';
 
 /**
  * [GET]
@@ -12,20 +10,4 @@ export const getCategoryIngredientList = async () => {
   );
 
   return response.data;
-};
-
-// 현재로서는 한번에 ingredient리스트를 전부 받아온다
-export const useCategoryIngredientList = () => {
-  const { data, ...rest } = useQuery(
-    ['search/get-category-ingredient'],
-    getCategoryIngredientList
-  );
-
-  // 우선 useMemo를 써서 memoization을 쓰긴 했는데, 꼭 필요한 부분인지는 고려 해봐야함.
-  const categoryIngredientList: CategoryIngredient[] = useMemo(
-    () => (data ? data : []),
-    [data]
-  );
-
-  return { list: categoryIngredientList, ...rest };
 };
