+++ types/ingredient.ts
@@ -3,7 +3,7 @@ type Ingredient = string;
 
 // 카테고리 재료
 type CategoryIngredient = {
-  category_name: string;
+  categoryName: string;
   ingredientNameList: Ingredient[];
 };
